Information and effect size for Weighted Log-rank test

Based on piecewise enrollment rate, failure rate, and dropout rates computes approximate information and effect size using an average hazard ratio model.

gs_info_wlr(
  enrollRates = tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6,
    9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  ratio = 1,
  events = NULL,
  analysisTimes = NULL,
  weight = wlr_weight_fh,
  approx = "asymptotic"
)

Arguments

enrollRates
enrollment rates

failRates
failure and dropout rates

ratio
Experimental:Control randomization ratio

events
Targeted minimum events at each analysis

analysisTimes
Targeted minimum study duration at each analysis

weight
weight of weighted log rank test * "1"=unweighted, * "n"=Gehan-Breslow, * "sqrtN"=Tarone-Ware, * "FH_p[a]_q[b]"= Fleming-Harrington with p=a and q=b

approx
approximate estimation method for Z statistics * "event driven" = only work under propotional hazard model with log rank test * "asymptotic"

Value

a tibble with columns Analysis, Time, N, Events, AHR, delta, sigma2, theta, info, info0. info, info0 contains statistical information under H1, H0, respectively. For analysis k, Time[k] is the maximum of analysisTimes[k] and the expected time required to accrue the targeted events[k]. AHR is expected average hazard ratio at each analysis.

Details

The AHR() function computes statistical information at targeted event times. The tEvents() function is used to get events and average HR at targeted analysisTimes.