Group sequential design power under non-proportional hazards

Group sequential design power under non-proportional hazards

Usage

gs_power_nph(
  enrollRates = tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6,
    9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  ratio = 1,
  events = c(30, 40, 50),
  analysisTimes = NULL,
  maxEvents = 45,
  upper = gs_b,
  upar = gsDesign(k = length(events), test.type = 1, n.I = events, maxn.IPlan =
    maxEvents, sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lower = gs_b,
  lpar = c(qnorm(0.1), rep(-Inf, 2)),
  r = 18
)

Arguments

enrollRates

enrollment rates

failRates

failure and dropout rates

ratio

Experimental:Control randomization ratio (not yet implemented)

events

Targeted events at each analysis

analysisTimes

Not yet implemented

maxEvents

Final planned events

upper

Function to compute upper bound

upar

Parameter passed to upper()

lower

Function to compute lower bound

lpar

Parameter passed to lower()

r

Control for grid size; normally leave at default of r=18

Value

a tibble with columns Analysis, Bound, Z, Probability, theta, Time, avehr, Events

Details

Need to be added

Specification

The contents of this section are shown in PDF user manual only.

Examples

library(gsDesign)
library(gsDesign2)
library(dplyr)

gs_power_nph() %>% filter(abs(Z) < Inf)
#> # A tibble: 4 × 10
#>   Analysis Bound     Z Probability  Time Events   AHR theta  info info0
#>      <int> <chr> <dbl>       <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1        1 Upper  2.51      0.0316  14.9   30.0 0.787 0.240  7.37  7.50
#> 2        2 Upper  2.15      0.113   19.2   40.0 0.744 0.295  9.79 10.0 
#> 3        3 Upper  2.12      0.192   24.5   50.0 0.713 0.339 12.2  12.5 
#> 4        1 Lower -1.28      0.0266  14.9   30.0 0.787 0.240  7.37  7.50