Group sequential design power using weighted log rank test under non-proportional hazards

Group sequential design power using weighted log rank test under non-proportional hazards

gs_power_wlr(
  enrollRates = tibble::tibble(Stratum = "All", duration = c(2, 2, 10), rate = c(3, 6,
    9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  ratio = 1,
  weight = wlr_weight_fh,
  approx = "asymptotic",
  events = c(30, 40, 50),
  analysisTimes = NULL,
  binding = FALSE,
  upper = gs_b,
  upar = gsDesign(k = length(events), test.type = 1, n.I = events, maxn.IPlan =
    max(events), sfu = sfLDOF, sfupar = NULL)$upper$bound,
  lower = gs_b,
  lpar = c(qnorm(0.1), rep(-Inf, length(events) - 1)),
  test_upper = TRUE,
  test_lower = TRUE,
  r = 18,
  tol = 1e-06
)

Arguments

enrollRates
enrollment rates

failRates
failure and dropout rates

ratio
Experimental:Control randomization ratio (not yet implemented)

weight
weight of weighted log rank test * "1"=unweighted, * "n"=Gehan-Breslow, * "sqrtN"=Tarone-Ware, * "FH_p[a]_q[b]"= Fleming-Harrington with p=a and q=b

approx
approximate estimation method for Z statistics * "event driven" = only work under propotional hazard model with log rank test * "asymptotic"

events
Targeted events at each analysis

analysisTimes
Minimum time of analysis

binding
indicator of whether futility bound is binding; default of FALSE is recommended

upper
Function to compute upper bound

upar
Parameter passed to upper()

lower
Function to compute lower bound

lpar
Parameter passed to lower()

test_upper
indicator of which analyses should include an upper (efficacy) bound; single value of TRUE (default) indicates all analyses; otherwise, a logical vector of the same length as info should indicate which analyses will have an efficacy bound

test_lower
indicator of which analyses should include an lower bound; single value of TRUE (default) indicates all analyses; single value FALSE indicated no lower bound; otherwise, a logical vector of the same length as info should indicate which analyses will have a lower bound

r
Integer, at least 2; default of 18 recommended by Jennison and Turnbull

tol
Tolerance parameter for boundary convergence (on Z-scale)

Specification

The contents of this section are shown in PDF user manual only.