MaxCombo Group sequential boundary crossing probabilities

MaxCombo Group sequential boundary crossing probabilities

gs_prob_combo(
  upper_bound,
  lower_bound,
  analysis,
  theta,
  corr,
  algorithm = GenzBretz(maxpts = 1e+05, abseps = 1e-05),
  ...
)

Arguments

upper_bound
a numeric vector of upper bound

lower_bound
a numeric vector of lower bound

analysis
an integer vector of the interim analysis index

theta
a numeric vector of effect size under alternative hypothesis

corr
a matrix of correlation matrix

algorithm
an object of class GenzBretz, Miwa or TVPACK specifying both the algorithm to be used as well as the associated hyper parameters.

...
additional parameters transfer to `mvtnorm::pmvnorm`